requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg[binary,pool]==3.1.19
alembic==1.13.2
pydantic==2.7.4
pydantic-settings==2.3.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
celery==5.3.6
redis==5.0.4
slowapi==0.1.9
pytest==8.2.2
httpx==0.27.0
